requirements.txt
astroid==2.11.6
autopep8==1.6.0
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
dill==0.3.5.1
flake8==4.0.1
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pylint==2.14.3
python-decouple==3.6
SQLAlchemy==1.4.37
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
Werkzeug==2.1.2
wrapt==1.14.1
